extract class for notification
