+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/UpdaterServiceDelegatee.scala
@@ -1,15 +1,10 @@
 package x7c1.linen.modern.init.updater
 
-import android.app.{NotificationManager, PendingIntent, Service}
-import android.content.Context.NOTIFICATION_SERVICE
+import android.app.Service
 import android.content.Intent
 import android.os.IBinder
-import android.support.v4.app.NotificationCompat
-import android.support.v4.content.LocalBroadcastManager
-import android.support.v7.app.NotificationCompat.Builder
-import x7c1.linen.glue.service.{ServiceControl, ServiceLabel}
+import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.modern.init.dev.DummyFactory
-import x7c1.linen.modern.init.updater.UpdaterServiceDelegatee.ActionTypeSample
 import x7c1.wheat.macros.intent.MethodCaller
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.service.CommandStartType
@@ -27,8 +22,6 @@ class UpdaterServiceDelegatee(service: Service with ServiceControl){
     None
   }
 
-  var notificationId: Int = 1
-
   def onStartCommand(intent: Intent, flags: Int, startId: Int): CommandStartType = {
     Log info s"[init] start:$startId, $intent"
 
@@ -39,45 +32,9 @@ class UpdaterServiceDelegatee(service: Service with ServiceControl){
     }
     NotSticky
   }
-  def showNotification() = {
-    val notificationIntent = new Intent(service, service getClassOf ServiceLabel.Updater)
-    val pendingIntent = PendingIntent.getService(service, 0, notificationIntent, PendingIntent.FLAG_IMMUTABLE)
-    val builder = new Builder(service).
-      setContentIntent(pendingIntent).
-      setContentTitle(s"sample title $notificationId").
-      setContentText(s"sample text $notificationId").
-      setProgress(100, notificationId, false).
-      setSmallIcon(android.R.drawable.ic_dialog_info)
-
-    val style0 = new NotificationCompat.InboxStyle(builder)
-      .setSummaryText(s"summary $notificationId")
-      .setBigContentTitle(s"title $notificationId")
-
-    val style = (1 to notificationId).foldLeft(style0){
-      case (s, i) => s.addLine(s"line $i")
-    }
-    val notification = style.build()
-
-    val manager = service.getSystemService(NOTIFICATION_SERVICE).asInstanceOf[NotificationManager]
-
-    Log info s"[n-id] $notificationId"
-    service.startForeground(notificationId, notification)
-    manager.notify(notificationId, notification)
-
-    sendMessage(notificationId)
-
-    notificationId += 1
-  }
   def onDestroy(): Unit = {
     Log info "[init]"
   }
-  def sendMessage(n: Int) = {
-    Log info s"[init] $n"
-    val intent = new Intent(ActionTypeSample)
-    intent.putExtra("sample-message", s"hello!!! $n")
-    LocalBroadcastManager.getInstance(service).sendBroadcast(intent)
-  }
-
 }
 class UpdaterMethods(service: Service with ServiceControl, startId: Int){
 
@@ -85,64 +42,12 @@ class UpdaterMethods(service: Service with ServiceControl, startId: Int){
 
   def createDummies(max: Int): Unit = async {
     Log info "[init]"
+    val notifier = new UpdaterServiceNotifier(service, max)
     DummyFactory.createDummies0(service)(max){ n =>
-      notifyProgress(max, n)
+      notifier.notifyProgress(n)
     }
-    notifyDone(max, max)
+    notifier.notifyDone()
     service stopSelf startId
-  }
-  private def manager =
-    service.getSystemService(NOTIFICATION_SERVICE).asInstanceOf[NotificationManager]
-
-  private def notifyDone(max: Int, current: Int): Unit = {
-
-    val notificationIntent = new Intent(service, service getClassOf ServiceLabel.Updater)
-    val pendingIntent =
-      PendingIntent.getService(service, 0, notificationIntent, PendingIntent.FLAG_IMMUTABLE)
-
-    val builder = new Builder(service).
-      setContentIntent(pendingIntent).
-      setContentTitle(s"Progress $current/$max").
-      setContentText(s"inserting").
-      setProgress(max, current, false/*indeterminate*/).
-      setAutoCancel(true).
-      setSmallIcon(android.R.drawable.ic_dialog_info)
-
-    val style0 = new NotificationCompat.InboxStyle(builder)
-      .setSummaryText(s"inserted $current/$max")
-      .setBigContentTitle("Progress")
-
-    val notification = style0.build()
-
-    /*
-    rebuild notification for kitkat device
-     */
-    service stopForeground {
-      val removeNotification = true
-      removeNotification
-    }
-    manager.notify(123, notification)
-  }
-  private def notifyProgress(max: Int, current: Int): Unit = {
-    val notificationIntent = new Intent(service, service getClassOf ServiceLabel.Updater)
-    val pendingIntent =
-      PendingIntent.getService(service, 0, notificationIntent, PendingIntent.FLAG_IMMUTABLE)
-
-    val builder = new Builder(service).
-      setContentIntent(pendingIntent).
-      setContentTitle(s"Progress $current/$max").
-      setContentText(s"inserting").
-      setProgress(max, current, false/*indeterminate*/).
-      setSmallIcon(android.R.drawable.ic_dialog_info)
-
-    val style0 = new NotificationCompat.InboxStyle(builder)
-      .setSummaryText(s"inserted $current/$max")
-      .setBigContentTitle("Progress")
-
-    val notification = style0.build()
-
-    service.startForeground(123, notification)
-    manager.notify(123, notification)
   }
 
   def baz: Unit = {
